Some styling and basic layout

+++ src/client/src/components/DeleteEvent.js
@@ -1,6 +1,6 @@
 import React, {useEffect, useState} from 'react'
 import axios from 'axios'
-
+import '../styles/eventcontrolpanel.css'
 
 export default function DeleteEvent() {
   const [events, setEvents] = useState([])
@@ -18,19 +18,22 @@ export default function DeleteEvent() {
   }
 
   const handleSubmit = async (eventId) => {
-    const token = sessionStorage.getItem('token')
-    console.log("Submitting:", eventId)
+    let test = window.confirm("Are you sure you want to delete this event?")
+    if (test) {
+      const token = sessionStorage.getItem('token')
+      console.log("Submitting:", eventId)
 
-    await axios.delete("http://localhost:7002/api/events/delete", {
-      headers: {
-        'Authorization': token,
-        "eventId": eventId
-      }})
-    .then(res => {
-      fetchEvents()
-      console.log(res)
-    })
-    .catch(err => console.log(err))
+      await axios.delete("http://localhost:7002/api/events/delete", {
+        headers: {
+          'Authorization': token,
+          "eventId": eventId
+        }})
+      .then(res => {
+        fetchEvents()
+        console.log(res)
+      })
+      .catch(err => console.log(err))
+    }
   }
 
   useEffect(() => {
@@ -39,9 +42,12 @@ export default function DeleteEvent() {
 
   return (
     <>
+    <p>Which event would you like to delete?</p>
+    <div className="event-buttons">
       {events ? (events.map((event) => {
         return (<button key={event._id} onClick={(e) => handleSubmit(event._id)}>{event.name}</button>)
       })): (null)}
+    </div>
     </>
   )
